Use `BigDecimal` to store coordinate

+++ src/models/location.rs
@@ -1,9 +1,10 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Location {
     pub id: i64,
     pub name: String,
-    pub longitude: f64,
-    pub latitude: f64,
+    pub longitude: BigDecimal,
+    pub latitude: BigDecimal,
 }
